Simplify the datatypes for the template expansion.
