+++ create-neon/data/templates/README.md.hbs
@@ -1,5 +1,5 @@
-# {{project.name}}
+# {{package.name}}
 
-{{#if project.description}}
-{{project.description.raw}}
+{{#if package.description}}
+{{package.description}}
 {{/if}}
